+++ packages/duckdb/src/tool.ts
@@ -1,4 +1,4 @@
-import { tool } from '@openassistant/utils';
+import { tool, generateId } from '@openassistant/utils';
 import { Table as ArrowTable, tableFromArrays } from 'apache-arrow';
 import { z } from 'zod';
 import { getDuckDB, QueryDuckDBFunctionContext } from './query';
@@ -9,18 +9,30 @@ import { QueryDuckDBComponent } from './queryTable';
  *
  * @example
  * ```typescript
- * import { localQuery } from '@openassistant/duckdb';
+ * import { getVercelAiTool } from '@openassistant/osm';
  *
- * const localQueryTool = {
- *   ...localQuery,
- *   context: {
- *     ...localQuery.context,
- *     getValues: (datasetName: string, variableName: string) => {
- *       // get the values of the variable from your dataset, e.g.
- *       return SAMPLE_DATASETS[datasetName].map((item) => item[variableName]);
- *     },
+ *
+ * // context
+ * const context = {
+ *   getValues: (datasetName: string, variableName: string) => {
+ *     // get the values of the variable from your dataset, e.g.
+ *     return SAMPLE_DATASETS[datasetName].map((item) => item[variableName]);
  *   },
  * }
+ *
+ * // onToolCompleted
+ * const onToolCompleted = (toolCallId: string, additionalData?: unknown) => {
+ *   // do something with the additionalData
+ * }
+ *
+ * // get the tool
+ * const localQueryTool = getVercelAiTool('localQuery', context, onToolCompleted);
+ *
+ * generateText({
+ *   model: 'gpt-4o-mini',
+ *   prompt: 'What are the venues in San Francisco?',
+ *   tools: {localQuery: localQueryTool},
+ * });
  * ```
  *
  * ### getValues()
@@ -40,8 +52,7 @@ import { QueryDuckDBComponent } from './queryTable';
  *
  */
 export const localQuery = tool({
-  description:
-    'You are a SQL (duckdb) expert. You can help to generate select query clause using the content of the dataset.',
+  description: `You are a SQL (duckdb) expert. You can help to query users datasets using select query clause.`,
   parameters: z.object({
     datasetName: z.string().describe('The name of the original dataset.'),
     variableNames: z
@@ -54,7 +65,9 @@ export const localQuery = tool({
       ),
     dbTableName: z
       .string()
-      .describe('The name of the table used in the sql string.'),
+      .describe(
+        'The name of the table used in the sql string. Please use the datasetName as the table name, and remove any space or special characters from the datasetName.'
+      ),
   }),
   execute: executeLocalQuery,
   context: {
@@ -82,33 +95,22 @@ async function executeLocalQuery(
   try {
     const { getValues, config, duckDB } =
       options.context as QueryDuckDBFunctionContext;
-    // if the variable names contain 'row_index', ignore it
-    // the row_index will be added later based on the columnData length
-    // the row_index is used to sync the selections of the query result table with the original dataset
-    const variableNamesWithoutRowIndex = variableNames.filter(
-      (name) => name !== 'row_index'
-    );
+
     // get values for each variable
-    const columnData = variableNamesWithoutRowIndex.reduce((acc, varName) => {
-      try {
-        acc[varName] = getValues(datasetName, varName);
-      } catch {
-        throw new Error(`variable ${varName} is not found in the dataset.`);
-      }
-      return acc;
-    }, {});
-    // add the row_index to the column data
-    columnData['row_index'] = Array.from(
-      { length: columnData[variableNamesWithoutRowIndex[0]].length },
-      (_, i) => i
-    );
+    const columnData = {};
+    for (const varName of variableNames) {
+      columnData[varName] = await getValues(datasetName, varName);
+    }
+
     // Create Arrow Table from column data with explicit type
     const arrowTable: ArrowTable = tableFromArrays(columnData);
+
     // Initialize DuckDB with external instance if provided
     const db = await getDuckDB(duckDB);
     if (!db) {
       throw new Error('DuckDB instance is not initialized');
     }
+
     const conn = await db.connect();
     const safeDbTableName = `${dbTableName}`;
     await conn.query(`DROP TABLE IF EXISTS ${safeDbTableName}`);
@@ -116,20 +118,39 @@ async function executeLocalQuery(
       name: safeDbTableName,
       create: true,
     });
+
     // query all table names in duckdb
     // const tableNamesResult = await conn.query('SHOW TABLES');
     // const tableNames = tableNamesResult.toArray().map(row => row.toJSON());
     // console.log(tableNames);
     const arrowResult = await conn.query(sql);
+
+    // here we don't pass the arrowResult to LLM or additionalData
+    // because the arrowResult is too large to be passed to LLM
+    // and we don't want to persist the arrowResult in store as well
     // const result = arrowResult.toArray().map((row) => row.toJSON());
+    // However, we still generate a filtered dataset name which could be used
+    // for possible tool callback to create a filtered dataset
+    const filteredDatasetName = `${datasetName}_filtered_${generateId()}`;
+
     await conn.close();
     // Get first 2 rows of the result as a json object
     const subResult = arrowResult.toArray().slice(0, 2);
-    const firstTwoRows = subResult.map((row) => row.toJSON());
+    const firstTwoRows = subResult.map((row) => {
+      const json = row.toJSON();
+      // Convert any BigInt values to strings
+      return Object.fromEntries(
+        Object.entries(json).map(([key, value]) => [
+          key,
+          typeof value === 'bigint' ? value.toString() : value
+        ])
+      );
+    });
 
     return {
       llmResult: {
         success: true,
+        filteredDatasetName,
         data: {
           firstTwoRows,
         },
@@ -141,6 +162,7 @@ async function executeLocalQuery(
         variableNames,
         datasetName,
         dbTableName,
+        filteredDatasetName,
         config,
       },
     };