+++ packages/echarts/src/register-tools.ts
@@ -34,7 +34,7 @@ export function getVercelAiTool(
   return getTool(tool, toolContext, onToolCompleted);
 }
 
-export function getVercelTools(
+export function getVercelAiTools(
   toolContext: ToolContext,
   onToolCompleted: OnToolCompleted
 ) {