+++ packages/geoda/src/utils.ts
@@ -5,6 +5,7 @@ import { getCachedWeightsById } from './weights/tool';
 import { SpatialToolContext } from './types';
 
 export function generateId() {
+  // generate a random id with 10 characters
   return Math.random().toString(36).substring(2, 15);
 }