+++ packages/keplergl/src/register-tools.ts
@@ -19,7 +19,7 @@ export function getVercelAiTool(
   return getTool(tool, toolContext, onToolCompleted);
 }
 
-export function getVercelTools(
+export function getVercelAiTools(
   toolContext: KeplerglToolContext,
   onToolCompleted: OnToolCompleted
 ) {