+++ packages/osm/src/register-tools.ts
@@ -43,7 +43,7 @@ export function getVercelAiTool(
   return getTool(tool, toolContext, onToolCompleted);
 }
 
-export function getVercelTools(
+export function getVercelAiTools(
   toolContext: OsmToolContext,
   onToolCompleted: OnToolCompleted
 ) {