+++ packages/common/CHANGELOG.md
@@ -3,6 +3,15 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.4.0](https://github.com/GeoDaCenter/openassistant/compare/@openassistant/common@0.0.5...@openassistant/common@0.4.0) (2025-05-07)
+
+### Features
+
+* allow using tools directly with vercel ai sdk ([#17](https://github.com/GeoDaCenter/openassistant/issues/17)) ([fef9fa2](https://github.com/GeoDaCenter/openassistant/commit/fef9fa2d21938c304a4415c7b2fc3972280e0908))
+* update playground ([#15](https://github.com/GeoDaCenter/openassistant/issues/15)) ([6ed4ca9](https://github.com/GeoDaCenter/openassistant/commit/6ed4ca97901626b84eb5667bdbab3d38cc33570a))
+
+## 0.0.6 (2025-01-21)
+
 ## [0.3.10](https://github.com/GeoDaCenter/openassistant/compare/@openassistant/common@0.0.5...@openassistant/common@0.3.10) (2025-05-06)
 
 ## 0.3.9 (2025-05-06)